Add page home show product

+++ db/migrate/20241016091619_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[6.0]
     create_table :products do |t|
       t.string :name
       t.text :description
-      t.decimal :price
+      t.bigint :price
       t.string :image
 
       t.timestamps
