+++ db/migrate/20241017040852_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.references :user, null: false, foreign_key: true
-      t.decimal :total_price
+      t.bigint :total_price
 
       t.timestamps
     end
